[refactor] exception ENUM화에 따른 도메인 코드 수정
