+++ src/main/java/kr/hhplus/be/server/domain/BusinessException.java
@@ -1,7 +1,5 @@
 package kr.hhplus.be.server.domain;
 
-import kr.hhplus.be.server.domain.point.PointErrorCode;
-
 public class BusinessException extends RuntimeException {
     private final ErrorCode errorCode;
