--- src/main/java/kr/hhplus/be/server/domain/model/PopularProduct.java
@@ -1,16 +1,0 @@
-package kr.hhplus.be.server.domain.model;
-
-public class PopularProduct {
-    private int productId;
-    private String name;
-    private int salesQuantity;
-    private double revenue;
-
-    public PopularProduct(int productId, String name, int salesQuantity, double revenue) {
-        this.productId = productId;
-        this.name = name;
-        this.salesQuantity = salesQuantity;
-        this.revenue = revenue;
-    }
-
-}
