--- src/main/java/kr/hhplus/be/server/domain/model/Product.java
@@ -1,19 +1,0 @@
-package kr.hhplus.be.server.domain.model;
-
-import lombok.Getter;
-
-@Getter
-public class Product {
-    private int id;
-    private String name;
-    private double price;
-    private int stock;
-
-    public Product(int id, String name, double price, int stock) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-        this.stock = stock;
-    }
-
-}
