+++ src/main/java/kr/hhplus/be/server/domain/point/enums/PointHistoryType.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.point;
+package kr.hhplus.be.server.domain.point.enums;
 
 public enum PointHistoryType {
     CHARGE,
