+++ src/main/java/kr/hhplus/be/server/domain/product/ProductStock.java
@@ -1,6 +1,7 @@
 package kr.hhplus.be.server.domain.product;
 
 import kr.hhplus.be.server.domain.BusinessException;
+import kr.hhplus.be.server.domain.product.enums.ProductErrorCode;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
