+++ src/main/java/kr/hhplus/be/server/domain/product/enums/ProductErrorCode.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.product;
+package kr.hhplus.be.server.domain.product.enums;
 
 import kr.hhplus.be.server.domain.ErrorCode;
