+++ src/main/java/kr/hhplus/be/server/domain/product/enums/SellingStatus.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.product;
+package kr.hhplus.be.server.domain.product.enums;
 
 public enum SellingStatus {
     SELLING,
