+++ src/main/java/kr/hhplus/be/server/domain/user/User.java
@@ -1,0 +1,24 @@
+package kr.hhplus.be.server.domain.user;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class User {
+    private Long id;
+    private String name;
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
+
+    public static User of(Long id, String name) {
+        if (name == null || name.isBlank()) {
+            throw new IllegalArgumentException("사용자 이름은 비어 있을 수 없습니다.");
+        }
+        LocalDateTime now = LocalDateTime.now();
+        return new User(id, name, now, now);
+    }
+}
